[refactor] #55 문제 도메인 enum 패키지 분리
